+++ test/FunctionMetadataGeneratorTests/FunctionMetadataGeneratorTests.cs
@@ -11,6 +11,7 @@ using System.Linq;
 using System.Reflection;
 using System.Threading.Tasks;
 using Azure.Messaging.EventHubs;
+using Azure.Messaging.ServiceBus;
 using Azure.Storage.Blobs;
 using Azure.Storage.Queues.Models;
 using Microsoft.Azure.Functions.Tests;
@@ -895,6 +896,59 @@ namespace Microsoft.Azure.Functions.SdkTests
         }
 
         [Fact]
+        public void ServiceBus_SDKTypeBindings()
+        {
+            var generator = new FunctionMetadataGenerator();
+            var module = ModuleDefinition.ReadModule(_thisAssembly.Location);
+            var typeDef = TestUtility.GetTypeDefinition(typeof(SDKTypeBindings_ServiceBus));
+            var functions = generator.GenerateFunctionMetadata(typeDef);
+            var extensions = generator.Extensions;
+
+            Assert.Equal(2, functions.Count());
+
+            AssertDictionary(extensions, new Dictionary<string, string>
+            {
+                { "Microsoft.Azure.WebJobs.Extensions.ServiceBus", "5.11.0" },
+            });
+
+            var serviceBusTriggerFunction = functions.Single(p => p.Name == nameof(SDKTypeBindings_ServiceBus.ServiceBusTriggerFunction));
+
+            ValidateFunction(serviceBusTriggerFunction, nameof(SDKTypeBindings_ServiceBus.ServiceBusTriggerFunction), GetEntryPoint(nameof(SDKTypeBindings_ServiceBus), nameof(SDKTypeBindings_ServiceBus.ServiceBusTriggerFunction)),
+                ValidateServiceBusTrigger);
+
+            var serviceBusBatchTriggerFunction = functions.Single(p => p.Name == nameof(SDKTypeBindings_ServiceBus.ServiceBusBatchTriggerFunction));
+
+            ValidateFunction(serviceBusBatchTriggerFunction, nameof(SDKTypeBindings_ServiceBus.ServiceBusBatchTriggerFunction), GetEntryPoint(nameof(SDKTypeBindings_ServiceBus), nameof(SDKTypeBindings_ServiceBus.ServiceBusBatchTriggerFunction)),
+                ValidateServiceBusBatchTrigger);
+
+            void ValidateServiceBusTrigger(ExpandoObject b)
+            {
+                AssertExpandoObject(b, new Dictionary<string, object>
+                {
+                    { "Name", "message" },
+                    { "Type", "serviceBusTrigger" },
+                    { "Direction", "In" },
+                    { "queueName", "queue" },
+                    { "Cardinality", "One" },
+                    { "Properties", new Dictionary<String, Object>( ) { { "SupportsDeferredBinding" , "True"} } }
+                });
+            }
+
+            void ValidateServiceBusBatchTrigger(ExpandoObject b)
+            {
+                AssertExpandoObject(b, new Dictionary<string, object>
+                {
+                    { "Name", "messages" },
+                    { "Type", "serviceBusTrigger" },
+                    { "Direction", "In" },
+                    { "queueName", "queue" },
+                    { "Cardinality", "Many" },
+                    { "Properties", new Dictionary<String, Object>( ) { { "SupportsDeferredBinding" , "True"} } }
+                });
+            }
+        }
+
+        [Fact]
         public void EventHubs_SDKTypeBindings()
         {
             var generator = new FunctionMetadataGenerator();
@@ -1158,11 +1212,27 @@ namespace Microsoft.Azure.Functions.SdkTests
                 throw new NotImplementedException();
             }
 
-
             [Function(nameof(QueueBinaryDataTrigger))]
             public static void QueueBinaryDataTrigger(
                 [QueueTrigger("queue")] BinaryData message)
 
+            {
+                throw new NotImplementedException();
+            }
+        }
+
+        private class SDKTypeBindings_ServiceBus
+        {
+            [Function(nameof(ServiceBusTriggerFunction))]
+            public static void ServiceBusTriggerFunction(
+                [ServiceBusTrigger("queue")] ServiceBusReceivedMessage message)
+            {
+                throw new NotImplementedException();
+            }
+
+            [Function(nameof(ServiceBusBatchTriggerFunction))]
+            public static void ServiceBusBatchTriggerFunction(
+                [ServiceBusTrigger("queue", IsBatched = true)] ServiceBusReceivedMessage[] messages)
             {
                 throw new NotImplementedException();
             }