+++ extensions/Worker.Extensions.ServiceBus/release_notes.md
@@ -4,6 +4,8 @@
 - My change description (#PR/#issue)
 -->
 
-### Microsoft.Azure.Functions.Worker.Extensions.ServiceBus <version>
+### Microsoft.Azure.Functions.Worker.Extensions.ServiceBus 5.10.0-preview3
 
-- <event>
+- Support binding to ServiceBusReceivedMessage (#1313)
+- Make Constants class internal (#1609)
+- Return ConversionResult.Failed when the wrong content-type or binding source is used (#1609)